requirements.txt
playsound==1.2.2
diffusers==0.18.1
googletrans==4.0.0-rc1
openai==0.27.8
gTTS==2.3.2
transformers==4.30.2
moviepy
opencv-python
pillow
pydub
